cart, orders and order_details related functionalities added

+++ backend/Server1/spring_boot_backend_security_jwt_template/src/main/java/com/app/POJOS/Product.java
@@ -1,18 +1,16 @@
 package com.app.POJOS;
 
+import java.util.List;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 import org.hibernate.validator.constraints.Range;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonProperty.Access;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -49,7 +47,7 @@ public class Product extends BaseEntity{
 	//
 	@Column
 	@Range(min=1)
-	private int quantity;
+	private Long quantity;
 	
 	//
 	@Enumerated(EnumType.STRING)
